render redirect while scraping
