+++ views/index.handlebars
@@ -1,0 +1,1 @@
+<h1>Hello!</h1>
