+++ views/layouts/main.handlebars
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width">
+    <!-- Compiled and minified CSS -->
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/0.97.8/css/materialize.min.css">
+    <!-- Google Icons -->
+     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
+    <!-- Local style sheet -->
+    <!-- <link rel="stylesheet" href="style/style.css" media="screen" title="no title"> -->
+    <title>Design Hoarder</title>
+  </head>
+  <body>
+    {{{ body }}}
+  </body>
+</html>
